LOGIC: import/export improvements

+++ src/main/java/cz/muni/fi/pv168/project/todoapp/business/service/export/GenericImportService.java
@@ -10,47 +10,59 @@ import cz.muni.fi.pv168.project.todoapp.business.service.export.batch.BatchImpor
 import cz.muni.fi.pv168.project.todoapp.business.service.export.batch.BatchOperationException;
 import cz.muni.fi.pv168.project.todoapp.business.service.export.format.Format;
 import cz.muni.fi.pv168.project.todoapp.business.service.export.format.FormatMapping;
+import cz.muni.fi.pv168.project.todoapp.storage.sql.db.TransactionExecutor;
+import cz.muni.fi.pv168.project.todoapp.ui.util.ImportOption;
 
+import javax.swing.*;
 import java.util.Collection;
 
 /**
  * Generic synchronous implementation of the {@link ImportService}.
  */
 public class GenericImportService implements ImportService {
-
     private final CrudService<Event> eventCrudService;
     private final CrudService<Category> categoryCrudService;
     private final CrudService<Template> templateCrudService;
     private final CrudService<Interval> intervalCrudService;
     private final FormatMapping<BatchImporter> importers;
+    private final TransactionExecutor transactionExecutor;
 
     public GenericImportService(CrudService<Event> eventCrudService,
                                 CrudService<Category> categoryCrudService,
                                 CrudService<Template> templateCrudService,
                                 CrudService<Interval> intervalCrudService,
-                                Collection<BatchImporter> importers) {
+                                Collection<BatchImporter> importers,
+                                TransactionExecutor transactionExecutor) {
         this.eventCrudService = eventCrudService;
         this.categoryCrudService = categoryCrudService;
         this.templateCrudService = templateCrudService;
         this.intervalCrudService = intervalCrudService;
         this.importers = new FormatMapping<>(importers);
+        this.transactionExecutor = transactionExecutor;
     }
 
     @Override
-    public void importData(String filePath) {
-        // TODO whole function should be transactional
-        eventCrudService.deleteAll();
-        templateCrudService.deleteAll();
-        categoryCrudService.deleteAll();
-        intervalCrudService.deleteAll();
-
+    public void importData(String filePath, ImportOption option) {
         Batch batch = getImporter(filePath).importBatch(filePath);
+        transactionExecutor.executeInTransaction(() -> {
+            if (option == ImportOption.REWRITE) {
+                eventCrudService.deleteAll();
+                templateCrudService.deleteAll();
+                categoryCrudService.deleteAll();
+                intervalCrudService.deleteAll();
 
-        // Categories have to be stored first - events and templates depends on them
-        batch.categories().forEach(this::createCategory);
-        batch.events().forEach(this::createEvent);
-        batch.templates().forEach(this::createTemplate);
-        batch.intervals().forEach(this::createInterval);
+                // Categories have to be stored first - events and templates depends on them
+                batch.categories().forEach(this::createCategory);
+                batch.events().forEach(this::createEvent);
+                batch.templates().forEach(this::createTemplate);
+                batch.intervals().forEach(this::createInterval);
+            } else {
+                batch.categories().forEach(this::mergeCategory);
+                batch.events().forEach(this::mergeEvent);
+                batch.templates().forEach(this::mergeTemplate);
+                batch.intervals().forEach(this::mergeInterval);
+            }
+        });
     }
 
     private void createEvent(Event event) {
@@ -67,6 +79,26 @@ public class GenericImportService implements ImportService {
 
     private void createInterval(Interval interval) {
         intervalCrudService.create(interval);
+    }
+
+    private void mergeEvent(Event event) {
+        if (eventCrudService.existsByGuid(event.getGuid())) eventCrudService.update(event);
+        else eventCrudService.create(event);
+    }
+
+    private void mergeCategory(Category category) {
+        if (categoryCrudService.existsByGuid(category.getGuid())) categoryCrudService.update(category);
+        else categoryCrudService.create(category);
+    }
+
+    private void mergeTemplate(Template template) {
+        if (templateCrudService.existsByGuid(template.getGuid())) templateCrudService.update(template);
+        else templateCrudService.create(template);
+    }
+
+    private void mergeInterval(Interval interval) {
+        if (intervalCrudService.existsByGuid(interval.getGuid())) intervalCrudService.update(interval);
+        else intervalCrudService.create(interval);
     }
 
     @Override
